backend/requirements.txt
Django==5.2.6
djangorestframework==3.15.2
django-cors-headers==4.4.0
google-generativeai==0.8.3
requests==2.31.0
beautifulsoup4==4.12.3
schedule==1.2.2
